fix(aws-amplify-react):Hide Forgot Password/Singup conditionally
The changes that were made to hide were partially reverted in #41d3184.
Updating code to include the fix
